Edit cfg file for ubuntu
`Dockerfile.ubuntu` was edited to remove flag -vm6058 (if it exists in OpenDSS `fpc` configuration file) from options when
compiling Opendss using `fpc`. The `fpc` version in the Ubuntu app repo
is does not have that have an warning message 6058 and throws an error
if that option is passed to `fpc`.